Fix rounded corners issue on help button by destroying notices
Empty this.$element in notices when there are no elements to display
as just hiding with CSS makes other items in the group no the last
child and therefore breaks any CSS rounded corners rules.

As notices are only ever generated at load, destroying the tool
is not a problem.

Bug: T63575
Change-Id: I63a044ca63f61b976f57e34ce9537034e68a2614