[BrowserTest] This test is unreliable in any version of Chrome
The issue is that when running automatically, upon loading the
page to be edited a second and third time, the cursor ends up
in a random place within the existing text of the article.

The test expects the cursor to always be at the start of the
article text, and this happens in Firefox.

I did a cursory check and I could not reproduce the issue of random
cursor placement upon multiple edits to the same article. There
may be a bug here, but it is not trivial to reproduce.

So let's take this test out of the Jenkins builds in the service
of more reliable green tests.

Change-Id: Ie3dd76c5b40f3035d43e6f0a06327adb70900f60