VisualEditorDataModule: Use definition summary instead of time tracking
The current approach matched historical RL conventions, but was quite
fragile. It used memcached to track hashes of all (or most?) factors
that influence the generated script.

Now that RL uses version hashes instead of timestamp, just hash the
script directly. Currently ResourceLoader falls back to hashing
the generated timestamp. Let's bypass this.

As a bonus, this also gets rid of the getMessages() hack which was
causing all messages to load twice and the module to be almost
twice a big as it could be.

It loaded messages once as part of getScript, via ve.init.platform.addMessages.
And a second time via getMessages() for which mw.loader calls mw.messages.set.

Bug: T94074
Change-Id: Ib8e688bf3a9c151cbb18016e10230c3a4bab46b8