Make texts of edit tabs' tooltips more helpful to new users
We used to re-label the other edit tab but this fell away for Wikimedia users
when we switched from adding a new tab to replacing the old one via config;
over-riding 'tooltip-ca-edit' looks like the best way of doing this here. The
mention of wikitext means it's less of a shock to users.

Bug: T99271
Change-Id: I0cb871095c9e5e8ef5a3c73b3bb105079b451dec