Enable new module versioning for VisualEditorDataModule
This module effectively uses it already since this module doesn't
add meta data to getDefinitionSummary, but rather adds all of getScript()
into the hash, which is exactly what enableModuleContentVersion is for.

Using this has the advantage of being well-tested and better cached,
as right now the script has to be computed twice (once for the response,
and once for the version header), whereas enableModuleContentVersion
uses RL's in-process cache.

Bug: T102578
Change-Id: I42fb720988b4b4b4310bde238dc321b62866f827