Remove redundant VisualEditorDataModule::getGitHeadHash static cache
This method is only called once, in getScript(). There is no need to
indefinitely cache it in the class.

Besides, the GitInfo class already caches it in much the same way in its
own class.

Also move the class instantiation of GitInto into getScript() so that it is
deferred until getScript() is needed – which in turn is only called once and
will naturally gargage collect it as it's now a local variable.

Change-Id: I545baa04e9ff2333fbf2ca7d0473dab0cf9269fd