MWTemplatePage: Re-word the no-description text to reflect being always there
The <visualeditor-dialog-transclusion-no-template-description> message
is now always displayed in template transclusion dialog, both when
adding a new one and when editing existing one. I think it's best to
just remove the incorrect part, as the dialog provides sufficient
context for what is happening.

The $2 parameter is no longer used or documented, but is still
supported for compatibility with existing translations.

Bug: T87130
Change-Id: Ibaa06c3bb1e9afbf0005b1605236f4be8cb15251