Put jquery.uls.data dependency in the correct place
This module is (and always has been) referenced by
ve.init.mw.Platform, so ensure it is loaded before that.

Previously this was not a problem because it was inside a
function that was never run unless the language search widget
was invoked.

This probably makes a good case to not put these functions
in platform, but that is for another day.

Bug: T127736
Change-Id: I119b27a397b15628a8de5f0b32ae636bc73a8ba4