Avoid trying to load unregistered modules
Ideally we would get an exception when trying to load
a non-existent module, but there is no difference between
a genuinely missing modules, and one which just doesn't exist
on this target (e.g. a desktop-only module on mobile or vice-versa).

This stops mobile VE from crashing every time a desktop-only
module is added to VE plugins.

Change-Id: Ia9ac0dd21b26c54ad779c6f499e68786c34ec873