DesktopWikitextArticleTarget#switchToVisualEditor: Use correct revision ID property
So we ensure we send the actual revision ID, not potentially undefined, to the
method. If we send undefined, it'll give us an etag for revision number 0...
And that will cause issues when we later attempt (in a future editor switch) to
pass an etag for revision 0 along with the correct revision.

Bug: T143920
Change-Id: I0c27956abdd26ae4f5f1b91010a5eab8ee0354ab