build: Change pathPrefix for eg-iframe back from '../lib' to 'lib'
Follows-up 94a7e34. The main fix there was adding the promise to
the example execution. The path change was not needed, and actually
breaks the demo, again. Only the 'docs' directory is published.
Accessing ../lib results in 404 Not Found after syncing to doc.wikimedia.org.

This is why we run 'grunt copy:jsduck', which copies those files into
the docs/ directory.

Change-Id: I2ea89bfbd980910510cbd8639f871ec931a768f1