Ensure required modulestyles are loaded on read after save
We already ensure any new regular modules are loaded after saving,
and that any changes to jsconfigvars are applied. But we don't
currently ensure styles-only modules are loaded.

Live Preview in MediaWiki core has always done this properly, but
the VE implementation forgot to consider the other two properties.

Issue was mentioned in otherwise-unrelated T156699.

Change-Id: I2e9fedc6dffcac4385ebac27571cc3e4f6fb1542